services/product/requirements-dev.txt
coverage==7.8.2
fastapi==0.115.12
gunicorn==23.0.0
httpx==0.28.1
motor==3.7.1
pydantic-settings==2.9.1
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
python-slugify==8.0.4
uvicorn==0.34.3